lib::timing: extract audio timing to its own module
timing is pretty self contained for all potential audio server setups
